refactor(VCST-5378): move UCP helpers into shared checkout
